Add logging, layered architecture and correct status code in returns.
