+++ src/game_leader/server/service/service.py
@@ -6,7 +6,7 @@ import threading
 
 import mysql.connector.errors as db_errors
 
-from src.common.database import connect, perform_write_query
+from src.common.data.database import connect, perform_write_query
 
 db = ""
 robot1Ip = os.environ.get("ROBOT1IP")  # Set these variables in your configuration.
@@ -30,7 +30,7 @@ def new_game(game_information):
     """
     Create a new game and writes it to the database.
 
-    Send "start_game" post requests to all of the robots.
+    Send "start" post requests to all of the robots.
 
     :param: information of the new game.
     :return: all robot responses.
